chore: Move VariantGrid to templates
