+++ src/domain/products/product-form/sections/variants.tsx
@@ -8,7 +8,7 @@ import IconTooltip from "../../../../components/molecules/icon-tooltip"
 import Input from "../../../../components/molecules/input"
 import TagInput from "../../../../components/molecules/tag-input"
 import BodyCard from "../../../../components/organisms/body-card"
-import VariantGrid from "../../../../components/variant-grid"
+import VariantGrid from "../../../../components/templates/variant-grid"
 import useNotification from "../../../../hooks/use-notification"
 import { getErrorMessage } from "../../../../utils/error-messages"
 import { getCombinations } from "../../../../utils/get-combinations"
